backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
loguru==0.7.3
pydantic==2.12.3
pydantic-settings==2.7.1
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
websockets==15.0.1
email-validator==2.3.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
